+++ src/model/quote_requests.rs
@@ -115,7 +115,7 @@ pub struct BarsRequest {
 #[derive(Debug, Clone, Serialize, Default)]
 pub struct BarsByPageRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub symbol: Option<String>,
+    pub symbols: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub period: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]